type annotation

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -3,7 +3,7 @@ from typing import List, Union
 """ workig with type annotation"""
 
 
-def sum_mixed_list(mixd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mixd_lst: List[Union[int, float]]) -> float:
     """
     type annotation
     """
